pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "api_migration_env"
version = "1.0.0"
description = "RL environment for legacy API migration tasks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai"
]

# This is the corrected entry point for the validator
[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "server*"]